requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pyyaml>=6.0
graphviz>=0.20.1
